issue-1251 adding PasswordStrengthModelFactory
